Added Format Dictionary support
