+++ .test_data/json-stdin_test_1.json
@@ -4,4 +4,11 @@
 { "timestamp": "2020-01-01 11:00:00", "data": "test_data" }
 { "timestamp": "2020-01-01 01:00:00", "data": "test_data" }
 { "timestamp": "2020-01-01 05:00:00", "data": "test_data" }
+{ "timestamp": "2020-01-01 04:00:00", "data": "test_data" }
 { "timestamp": "2020-01-01 06:00:00", "data": "test_data" }
+{ "timestamp": "2020-01-01 00:00:00", "data": "test_data" }
+{ "timestamp": "2020-01-01 00:00:00", "data": "test_data" }
+{ "timestamp": "2020-01-01 00:00:00", "data": "test_data" }
+{ "timestamp": "2020-01-01 00:00:00", "data": "test_data" }
+{ "timestamp": "2020-01-01 00:00:00", "data": "test_data" }
+{ "timestamp": "2020-01-06 00:00:00", "data": "test_data" }
