+++ Cargo.toml
@@ -18,5 +18,6 @@ log = "0.4.14"
 regex = "1.4.3"
 serde = {version = "1", features = ["derive"]}
 serde_json = "1.0"
+serde_yaml = "0.8"
 simplelog = "0.9.0"
 structopt = "0.3.21"
